register, login, checking token and logout done

+++ model/index.js
@@ -2,15 +2,19 @@ const Contact = require('./schemas/contact')
 
 //for mongoost
 
-const listContacts = async () => {
-  const results = await Contact.find({})
+const listContacts = async (userId) => {
+  const results = await Contact.find({ owner: userId}).populate({
+    path: 'owner',
+    select: 'email subscription -_id'
+  })
   return results
 }
 
-const getContactById = async (contactId) => {
-  const result = await Contact.findOne({ _id: contactId })
-  console.log(result.id)
-  console.log(result._id)
+const getContactById = async (contactId, userId) => {
+  const result = await Contact.findOne({ _id: contactId, owner: userId }).populate({
+    path: 'owner',
+    select: 'email subscription -_id'
+  })
   return result
 }
 
@@ -19,18 +23,18 @@ const addContact = async (body) => {
   return result
 }
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (contactId, body, userId) => {
 
   const result = await Contact.findByIdAndUpdate(
-    { _id: contactId }, 
+    { _id: contactId, owner: userId }, 
     { ...body },
     { new: true}
     )
   return result
 } 
 
-const removeContact = async (contactId) => {
- const result = await Contact.findByIdAndRemove({_id: contactId})
+const removeContact = async (contactId, userId) => {
+ const result = await Contact.findByIdAndRemove({_id: contactId, owner: userId})
   return result
 }
